+++ lib/taro/types/base_type.rb
@@ -20,5 +20,6 @@ Taro::Types::BaseType = Struct.new(:object) do
   extend Taro::Types::Shared::OpenAPIName
   extend Taro::Types::Shared::OpenAPIType
   extend Taro::Types::Shared::Rendering
+  include Taro::Types::Shared::Caching
   include Taro::Types::Shared::Errors
 end